+++ src/test/kotlin/com/github/themartdev/intellijgleam/parser/FunctionTest.kt
@@ -18,4 +18,6 @@ class FunctionTest : GleamParsingTestCase() {
     fun testLabeledArguments() = assertParsedCorrectly()
 
     fun testPipelines() = assertParsedCorrectly()
+
+    fun testShorthandLabel() = assertParsedCorrectly()
 }